backend/requirements.txt
# Backend Requirements for Cotton Leaf Disease Detection API

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Deep Learning and Computer Vision
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.227
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3

# ONNX Runtime (optional, for ONNX inference)
onnxruntime-gpu==1.16.3
onnx==1.15.0

# Data processing
pyyaml==6.0.1
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# AI/LLM for recommendations
google-generativeai==0.3.2

# Development
pytest==7.4.3
httpx==0.25.1
